Initial commit which adds basic skeleton for operator
This PR add basic skeleton for kubevirt-tekton-tasks-operator.
Purpose of this operator is to deploy and take care of https://github.com/kubevirt/kubevirt-tekton-tasks
and example pipelines. Operator will be deployed as a part of HCO.
Operator will not be deployed by default. Users will have to set a
special bool attribute in HCO CR (e.g. deployTektonOperator). Deployment
of operator will be possible even after installation of cnv. In case a
user would like to use an operator, he/she has to install tekton
pipelines / openshift pipelines into cluster. If tekton is not
installed, the operator will be deployed and will be waiting for
installation of tekton and there will be a message of required crds in
the operator log. After tekton is installed, the operator will deploy
tasks and pipelines.

Signed-off-by: Karel Šimon <ksimon@redhat.com>