fix: revert kubevirt version back to v0.58.0
Kubevirt v0.58.1 and 0.59.0-rc.0 probably contains bug, which is causing
that windows vm is in pending state. The virt-launcher is in unschedulable
state (probably due to TSC node selector). Test lanes, which are testing
example pipelines are failing due to that issue. Reverting kubevirt
version back will fix the test lanes, until the issue is fixed in
kubevirt project.
Signed-off-by: Karel Šimon <ksimon@redhat.com>