fix: revert back the hardcoded version of kubevirt for tests fix: remove TSC labels from worker nodes in e2e tests
Latest kubevirt version introduced tsc node selectors for windows vms. This is causing problems during scheduling of windows VM, which are in pending state, due to different tsc labels on nodes. Currently there is no easy fix in kubevirt, because it would break migration flow. Due to that, we have to disable node-labeller and remove the tsc labels from nodes, so virt-controller can schedule the VMs on more nodes. This approach will break migration flow, but in our tests we do not test it.

Signed-off-by: Karel Šimon <ksimon@redhat.com>