fix: generate imported iso name
The imported DV is mounted to modify-windows-iso task. When
user tries to rerun the efi pipeline, pipeline will fail, because
PVC already exists even allowReplaceISODV param is set to true.
This is because already mounted PVC can't be deleted and kubernetes sets
status of PVC to terminating (but the pvc stays in cluster until
modify-windows-iso pod exists) when the task tries to delete it.
This PR update pipeline to create new DV/PVC for each pipelineRun and
then deletes it. The PVC will then be deleted when user deletes
pipelineRun.

Signed-off-by: Karel Šimon <ksimon@redhat.com>